requirements.txt
flask
langchain 
transformers
requests 
python-dotenv
